Working in code

+++ src/main/kotlin/me/gilberto/essentials/config/ConfigMain.kt
@@ -1,6 +1,8 @@
 package me.gilberto.essentials.config
 
 import me.gilberto.essentials.EssentialsInstance.instance
+import me.gilberto.essentials.management.Manager.consoleMessage
+import me.gilberto.essentials.management.Manager.disableplugin
 import me.gilberto.essentials.management.Manager.pluginpastedir
 import org.bukkit.configuration.file.YamlConfiguration
 import java.io.File
@@ -8,32 +10,67 @@ import java.nio.file.Files
 
 
 object ConfigMain {
+    var configs = HashSet<YamlConfiguration>()
     lateinit var econf: YamlConfiguration
     fun startconfig() {
-        copyconfig("GD_EssentialsConfig")
-        econf = YamlConfiguration.loadConfiguration(File(pluginpastedir(), "GD_EssentialsConfig.yml"))
+        econf = copyconfig("GD_EssentialsConfig") ?: return
         ConfigLoad()
     }
+    fun addtoconfig (source: YamlConfiguration, path: String, value: Any) {
+        source.addDefault(path, value)
+        reloadconfig()
+    }
 
-    private fun copyconfig(source: String) {
+    fun reloadconfig() {
+        fun test(source: YamlConfiguration) {
+            val check: YamlConfiguration
+            try {
+                check = YamlConfiguration.loadConfiguration(File(pluginpastedir(), "${source.name}.yml"))
+            } catch (Ex: Exception) {
+                Ex.printStackTrace()
+                consoleMessage("§cProblema na config $source, retornando a antiga!")
+                source.save(source.currentPath)
+                return
+            }
+            check.save(source.currentPath)
+        }
+        for (i in configs) {
+            test(i)
+        }
+    }
+
+    private fun copyconfig(source: String) : YamlConfiguration? {
         val configfile = File(pluginpastedir(), "$source.yml")
-        val resource = instance.javaClass.getResourceAsStream("/$source.yml") ?: return
+        val resource = instance.javaClass.getResourceAsStream("/$source.yml") ?: return YamlConfiguration.loadConfiguration(File(pluginpastedir(), "$source.yml"))
         if (configfile.exists()) {
-            val v = YamlConfiguration.loadConfiguration(File(pluginpastedir(), "$source.yml")).getDouble("Version-file")
+            val check : YamlConfiguration
+            try {
+                check = YamlConfiguration.loadConfiguration(File(pluginpastedir(), "$source.yml"))
+            }
+            catch (Ex: Exception) {
+                consoleMessage("§cProblema na config $source.")
+                disableplugin()
+                Ex.printStackTrace()
+                return null
+            }
+            val v = check.getDouble("Version-file")
             val checkfile = File(pluginpastedir(), "$source-check.yml")
+            if (checkfile.exists()) checkfile.delete()
             Files.copy(resource, checkfile.toPath())
             val vc = YamlConfiguration.loadConfiguration(checkfile).getDouble("Version-file")
             if (vc > v) {
-                ConfigVersionChecker(configfile, checkfile)
+                ConfigVersionChecker(configfile, checkfile, vc.toString())
             }
             else {
-                checkfile.delete()
+                ConfigChecker(configfile, checkfile)
             }
-            return
+            val b = YamlConfiguration.loadConfiguration(File(pluginpastedir(), "$source.yml"))
+            configs.add(b)
+            return b
         }
         else (File(pluginpastedir()).mkdirs())
         Files.copy(resource, configfile.toPath())
-        return
+        return YamlConfiguration.loadConfiguration(File(pluginpastedir(), "$source.yml"))
     }
 
     fun getString(source: YamlConfiguration, path: String): String = source.getString(path)
